+++ src/frontend/apps/impress/src/features/docs/doc-export/inline-content-mapping/index.ts
@@ -1,2 +1,3 @@
 export * from './interlinkingLinkPDF';
 export * from './interlinkingLinkDocx';
+export * from './interlinkingLinkODT';