+++ src/frontend/apps/impress/src/features/docs/doc-export/types.ts
@@ -51,3 +51,13 @@ export type DocsExporterDocx = Exporter<
   IRunPropertiesOptions,
   TextRun
 >;
+
+export type DocsExporterODT = Exporter<
+  DocsBlockSchema,
+  DocsInlineContentSchema,
+  DocsStyleSchema,
+  React.ReactNode,
+  React.ReactNode,
+  Record<string, string>,
+  React.ReactNode
+>;